Split LoginInfo into header and footer
