+++ frontend/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <meta charset="UTF-8" name="viewport" content="width=device-width, initial-scale=1" />
+    <meta charset="UTF-8" name="viewport" content="width=300, maximum-scale=1" />
     <title>Next Generation Scrum Poker</title>
   </head>
   <body>
