+++ frontend/src/Components/Header.tsx
@@ -1,0 +1,11 @@
+import tngLogo from '../img/tng.svg';
+import classes from './Header.module.css';
+
+export const Header = () => (
+  <header className={classes.header}>
+    NEXT&nbsp;GENERATION SCRUM&nbsp;POKER
+    <a href="https://www.tngtech.com/" target="_blank">
+      <img src={tngLogo} alt="TNG Logo" className={classes.logo} />
+    </a>
+  </header>
+);
